verbs: Move fabric, domain, and MR calls into new file
Add a verbs_domain.c file and place the related fabric,
domain, and MR calls there.  There isn't enough functionality
to bother separating these into 3 files.
