+++ prov/verbs/src/fi_verbs.c
@@ -37,8 +37,6 @@
 #include "fi_verbs.h"
 
 
-static int fi_ibv_fabric(struct fi_fabric_attr *attr, struct fid_fabric **fabric,
-			 void *context);
 static void fi_ibv_fini(void);
 
 static const char *local_node = "localhost";
@@ -193,235 +191,6 @@ ssize_t fi_ibv_send_iov_flags(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr,
 	wr->send_flags = VERBS_INJECT_FLAGS(ep, len, flags) | VERBS_COMP_FLAGS(ep, flags);
 
 	return fi_ibv_send(ep, wr, len, count, context);
-}
-
-static int fi_ibv_mr_close(fid_t fid)
-{
-	struct fi_ibv_mem_desc *mr;
-	int ret;
-
-	mr = container_of(fid, struct fi_ibv_mem_desc, mr_fid.fid);
-	ret = -ibv_dereg_mr(mr->mr);
-	if (!ret)
-		free(mr);
-	return ret;
-}
-
-static struct fi_ops fi_ibv_mr_ops = {
-	.size = sizeof(struct fi_ops),
-	.close = fi_ibv_mr_close,
-	.bind = fi_no_bind,
-	.control = fi_no_control,
-	.ops_open = fi_no_ops_open,
-};
-
-static int
-fi_ibv_mr_reg(struct fid *fid, const void *buf, size_t len,
-	   uint64_t access, uint64_t offset, uint64_t requested_key,
-	   uint64_t flags, struct fid_mr **mr, void *context)
-{
-	struct fi_ibv_mem_desc *md;
-	int fi_ibv_access;
-	struct fid_domain *domain;
-
-	if (flags)
-		return -FI_EBADFLAGS;
-
-	if (fid->fclass != FI_CLASS_DOMAIN) {
-		return -FI_EINVAL;
-	}
-	domain = container_of(fid, struct fid_domain, fid);
-
-	md = calloc(1, sizeof *md);
-	if (!md)
-		return -FI_ENOMEM;
-
-	md->domain = container_of(domain, struct fi_ibv_domain, domain_fid);
-	md->mr_fid.fid.fclass = FI_CLASS_MR;
-	md->mr_fid.fid.context = context;
-	md->mr_fid.fid.ops = &fi_ibv_mr_ops;
-
-	fi_ibv_access = IBV_ACCESS_LOCAL_WRITE;
-	if (access & FI_REMOTE_READ)
-		fi_ibv_access |= IBV_ACCESS_REMOTE_READ;
-	if (access & FI_REMOTE_WRITE)
-		fi_ibv_access |= IBV_ACCESS_REMOTE_WRITE | IBV_ACCESS_REMOTE_ATOMIC;
-
-	md->mr = ibv_reg_mr(md->domain->pd, (void *) buf, len, fi_ibv_access);
-	if (!md->mr)
-		goto err;
-
-	md->mr_fid.mem_desc = (void *) (uintptr_t) md->mr->lkey;
-	md->mr_fid.key = md->mr->rkey;
-	*mr = &md->mr_fid;
-	return 0;
-
-err:
-	free(md);
-	return -errno;
-}
-
-static int fi_ibv_close(fid_t fid)
-{
-	struct fi_ibv_domain *domain;
-	int ret;
-
-	domain = container_of(fid, struct fi_ibv_domain, domain_fid.fid);
-	if (domain->pd) {
-		ret = ibv_dealloc_pd(domain->pd);
-		if (ret)
-			return -ret;
-		domain->pd = NULL;
-	}
-
-	free(domain);
-	return 0;
-}
-
-static int fi_ibv_open_device_by_name(struct fi_ibv_domain *domain, const char *name)
-{
-	struct ibv_context **dev_list;
-	int i, ret = -FI_ENODEV;
-
-	if (!name)
-		return -FI_EINVAL;
-
-	dev_list = rdma_get_devices(NULL);
-	if (!dev_list)
-		return -errno;
-
-	for (i = 0; dev_list[i]; i++) {
-		if (!strcmp(name, ibv_get_device_name(dev_list[i]->device))) {
-			domain->verbs = dev_list[i];
-			ret = 0;
-			break;
-		}
-	}
-	rdma_free_devices(dev_list);
-	return ret;
-}
-
-static struct fi_ops fi_ibv_fid_ops = {
-	.size = sizeof(struct fi_ops),
-	.close = fi_ibv_close,
-	.bind = fi_no_bind,
-	.control = fi_no_control,
-	.ops_open = fi_no_ops_open,
-};
-
-static struct fi_ops_mr fi_ibv_domain_mr_ops = {
-	.size = sizeof(struct fi_ops_mr),
-	.reg = fi_ibv_mr_reg,
-	.regv = fi_no_mr_regv,
-	.regattr = fi_no_mr_regattr,
-};
-
-static struct fi_ops_domain fi_ibv_domain_ops = {
-	.size = sizeof(struct fi_ops_domain),
-	.av_open = fi_no_av_open,
-	.cq_open = fi_ibv_cq_open,
-	.endpoint = fi_ibv_open_ep,
-	.scalable_ep = fi_no_scalable_ep,
-	.cntr_open = fi_no_cntr_open,
-	.poll_open = fi_no_poll_open,
-	.stx_ctx = fi_no_stx_context,
-	.srx_ctx = fi_no_srx_context,
-};
-
-static int
-fi_ibv_domain(struct fid_fabric *fabric, struct fi_info *info,
-	   struct fid_domain **domain, void *context)
-{
-	struct fi_ibv_domain *_domain;
-	struct fi_info *fi;
-	int ret;
-
-	fi = fi_ibv_search_verbs_info(NULL, info->domain_attr->name);
-	if (!fi)
-		return -FI_EINVAL;
-
-	ret = fi_ibv_check_domain_attr(info->domain_attr, fi);
-	if (ret)
-		return ret;
-
-	_domain = calloc(1, sizeof *_domain);
-	if (!_domain)
-		return -FI_ENOMEM;
-
-	ret = fi_ibv_open_device_by_name(_domain, info->domain_attr->name);
-	if (ret)
-		goto err;
-
-	_domain->pd = ibv_alloc_pd(_domain->verbs);
-	if (!_domain->pd) {
-		ret = -errno;
-		goto err;
-	}
-
-	_domain->domain_fid.fid.fclass = FI_CLASS_DOMAIN;
-	_domain->domain_fid.fid.context = context;
-	_domain->domain_fid.fid.ops = &fi_ibv_fid_ops;
-	_domain->domain_fid.ops = &fi_ibv_domain_ops;
-	_domain->domain_fid.mr = &fi_ibv_domain_mr_ops;
-
-	*domain = &_domain->domain_fid;
-	return 0;
-err:
-	free(_domain);
-	return ret;
-}
-
-static int fi_ibv_fabric_close(fid_t fid)
-{
-	free(fid);
-	return 0;
-}
-
-static struct fi_ops fi_ibv_fi_ops = {
-	.size = sizeof(struct fi_ops),
-	.close = fi_ibv_fabric_close,
-	.bind = fi_no_bind,
-	.control = fi_no_control,
-	.ops_open = fi_no_ops_open,
-};
-
-static struct fi_ops_fabric fi_ibv_ops_fabric = {
-	.size = sizeof(struct fi_ops_fabric),
-	.domain = fi_ibv_domain,
-	.passive_ep = fi_ibv_passive_ep,
-	.eq_open = fi_ibv_eq_open,
-	.wait_open = fi_no_wait_open,
-};
-
-static int fi_ibv_fabric(struct fi_fabric_attr *attr, struct fid_fabric **fabric,
-			 void *context)
-{
-	struct fi_ibv_fabric *fab;
-	struct fi_info *info;
-	int ret;
-
-	ret = fi_ibv_init_info();
-	if (ret)
-		return ret;
-
-	info = fi_ibv_search_verbs_info(attr->name, NULL);
-	if (!info)
-		return -FI_ENODATA;
-
-	ret = fi_ibv_check_fabric_attr(attr, info);
-	if (ret)
-		return -FI_ENODATA;
-
-	fab = calloc(1, sizeof(*fab));
-	if (!fab)
-		return -FI_ENOMEM;
-
-	fab->fabric_fid.fid.fclass = FI_CLASS_FABRIC;
-	fab->fabric_fid.fid.context = context;
-	fab->fabric_fid.fid.ops = &fi_ibv_fi_ops;
-	fab->fabric_fid.ops = &fi_ibv_ops_fabric;
-	*fabric = &fab->fabric_fid;
-	return 0;
 }
 
 static void fi_ibv_fini(void)
